+++ src/Concordium/Client/Types/Contract/Schema.hs
@@ -122,6 +122,8 @@ data SchemaType =
   | Struct Fields
   | Enum [(Text, Fields)]
   | String SizeLength
+  | UInt128
+  | Int128
   | ContractName SizeLength
   | ReceiveName SizeLength
   deriving (Eq, Generic, Show)
@@ -165,6 +167,8 @@ instance S.Serialize SchemaType where
       20 -> S.label "Struct" $ Struct <$> S.get
       21 -> S.label "Enum"   $ Enum <$> getListOfWithSizeLen Four (S.getTwoOf getText S.get)
       22 -> S.label "String" $ String <$> S.get
+      23 -> S.label "UInt128" $ pure UInt128
+      24 -> S.label "Int128"  $ pure Int128
       25 -> S.label "ContractName" $ ContractName <$> S.get
       26 -> S.label "ReceiveName"  $ ReceiveName <$> S.get
       x  -> fail [i|Invalid SchemaType tag: #{x}|]
@@ -193,6 +197,8 @@ instance S.Serialize SchemaType where
     Struct fields -> S.putWord8 20 <> S.put fields
     Enum enum     -> S.putWord8 21 <> putListOfWithSizeLen Four (S.putTwoOf putText S.put) enum
     String sl     -> S.putWord8 22 <> S.put sl
+    UInt128 -> S.putWord8 23
+    Int128  -> S.putWord8 24
     ContractName sl     -> S.putWord8 25 <> S.put sl
     ReceiveName sl      -> S.putWord8 26 <> S.put sl