+++ test/Spec.hs
@@ -2,6 +2,7 @@ module Main where
 
 import Test.Hspec
 
+import qualified DataTests.DoubleWordSpec
 import qualified SimpleClientTests.CliSpec
 import qualified SimpleClientTests.EncryptionSpec
 
@@ -9,3 +10,4 @@ main :: IO ()
 main = hspec $ parallel $ do
   SimpleClientTests.CliSpec.tests
   SimpleClientTests.EncryptionSpec.tests
+  DataTests.DoubleWordSpec.tests