+++ ChangeLog.md
@@ -2,3 +2,4 @@
 
 ## Unreleased changes
 - Add ContractName and ReceiveName to schema.
+- Add UInt128 and Int128 to schema type